feat: refactor sync
